Fixes formatting in ssl full path default var
The changes in abdb5d2f21aec60a8bb2f5fd148cb7e518c5a4e1 introduced a
folding scalar operator in the YAML vars file, but inappropriately
retained the double quotes surrounding the value. The folding operator
makes the string content literal, so the double quotes are no longer
required, and in fact break the template writing. Fortunately the
validate parameter on the template tasks for Topbeat and Filebeat
prevent a broken config from being written to a system.