Stripped out custom SSL logic
Should assume that user will take advantage of another role to handle
ssl cert distribution. When crusing through the documentation also noticed
that by default the filebeats will utilize the system CA store so it is
possible to enable encryption AND not pass in a custom CA. That is now
reflected in the variables as well.