Modify logserver reference variable to drop ip ref
This could have been particularly confusing because you can easily use a DNS
name in this field. In particularly situations where you are using SSL, and the
logserver cert is pegged to a hostname, you'll need to specify a hostname here.
Removing the `ip_address` from the variable therefore makes it more obvious that
this field will gladly take an IP address or DNS name