Hook-up E2E testing
Bring in Logstash and Elasticsearch containers and chain them together. Once the
role is deployed to the base container, the agents will begin to write out to
logstash, which immediately dumps to elasticsearch. Then using testinfra, I am
ensuring that the index got written and that there are is at least more than 0
hits of entries from the agents. Not the most elegant solution but it does the
job.