Remove CI/molecule tests :(
This repository isn't getting a lot of love and the tests are failing
over. I made a new ticket to circle back and update them, but for now I
highly recommend we just rip them out and re-vive them in a separate
process.

(Not only are they failing but the requirements files are triggering
security github alerts)